backend/requirements.txt
# Core Django
Django>=5.2.4
djangorestframework>=3.15.2

# Production server
gunicorn>=21.2.0

# Database
psycopg2-binary>=2.9.7
dj-database-url>=2.1.0

# CORS and Security
django-cors-headers>=4.3.1
whitenoise>=6.6.0

# Configuration
python-decouple>=3.8

# Authentication and Authorization
django-allauth>=0.57.0
python-social-auth>=0.3.6
social-auth-app-django>=5.4.0
django-oauth-toolkit>=1.7.1
google-auth>=2.25.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0

# API and Filtering
django-filter>=23.3
django-extensions>=3.2.3

# Background Tasks
celery>=5.3.4
redis>=5.0.1

# File Handling
Pillow>=10.0.1
boto3>=1.34.0
django-storages>=1.14.2

# Payment Processing
stripe>=7.7.0

# Communication
twilio>=8.11.0

# Security and Cryptography
cryptography>=41.0.7
PyJWT>=2.8.0

# HTTP Requests
requests>=2.31.0

# Development Tools
black>=23.12.1
flake8>=6.1.0

# Testing
pytest-django>=4.7.0
factory-boy>=3.3.0

# Production
gunicorn>=21.2.0
